==> scATAC-seq_Buenrostro2018/2. scATAC-seq peaks, TSS annotation Buenrostro.ipynb <==
#### This section relies on the CellOracle package for annotating scATAC-seq peak calls with Transcription Start Site (TSS) information and the gene names associated with the TSS sites. For further information, please refer to the CellOracle documentation available at https://morris-lab.github.io/CellOracle.documentation/index.html.

### Import library

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os, sys, shutil, importlib, glob
from tqdm.notebook import tqdm
from celloracle import motif_analysis as ma
import celloracle as co
co.__version__

import warnings
import numba
# Code that triggers the warning

with warnings.catch_warnings():
    warnings.simplefilter("ignore", category=numba.NumbaDeprecationWarning)

# config InlineBackend.figure_format = 'retina'

plt.rcParams['figure.figsize'] = [6, 4.5]
plt.rcParams["savefig.dpi"] = 300

### Load data

# Load scATAC-seq peak list.
peaks = pd.read_csv("~/Desktop/Buenrostro/res_Buenrostro2018_all_peaks.csv", index_col=0)
peaks = peaks.x.values
print(peaks)

# Load Cicero coaccessibility scores.
cicero_connections = pd.read_csv("~/Desktop/Buenrostro/res_Buenrostro2018_cicero_connections.csv", index_col=0)
print(cicero_connections.head())

### TSS annotation

ma.SUPPORTED_REF_GENOME

##!! Please make sure to specify the correct reference genome here
tss_annotated = ma.get_tss_info(peak_str_list=peaks, ref_genome="hg19") 

# Check results
print(tss_annotated.tail())

### Integrate TSS info and cicero connections

# The resulting file obtained from the integration process consists of three columns: ["peak_id", "gene_short_name", "coaccess"].
# The "peak_id" column represents either the TSS peak or the peaks that are connected to a TSS peak.
# The "gene_short_name" column contains the gene name associated with the TSS site.
# The "coaccess" column indicates the coaccessibility score between the peak and a TSS peak. A score of 1 signifies that the peak itself is a TSS.

integrated = ma.integrate_tss_peak_with_cicero(tss_peak=tss_annotated, cicero_connections=cicero_connections)
print(integrated.shape)
print(integrated.head())

### Filter out peaks with low coaccessibility scores

peak = integrated[integrated.coaccess >= 0.8]
peak = peak[["peak_id", "gene_short_name"]].reset_index(drop=True)
print(peak.shape)
print(peak.head())

peak.to_csv("~/Desktop/Buenrostro/res_Buenrostro2018_processed_peak_file.csv")
